+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,14 @@
+<?php
+
+$lang = &$GLOBALS['TL_LANG']['tl_settings'];
+
+/**
+ * Fields
+ */
+$lang['activateDbCache'] = ['Database-Cache aktivieren', 'Wählen Sie diese Option, um das Modul zu aktivieren.'];
+$lang['dbCacheMaxTime']  = ['Maximale Cache-Zeit', 'Geben Sie hier die Zeit ein, die ein Wert im Cache verbleiben darf.'];
+
+/**
+ * Legends
+ */
+$lang['db_cache_legend'] = 'Database-Cache';